+++ NEWS.md
@@ -1,4 +1,4 @@
-# campfin (development version)
+# campfin 1.0.8
 
 * Add `string` argument to `guess_delim()` to read the input as a single line
   of a file regardless of `\n` presence.